完善BSP

+++ User/Boards/bsp_led.h
@@ -23,7 +23,7 @@ typedef enum {
 	BSP_LED6,
 	BSP_LED7,
 	BSP_LED8,
-#elif defined USE_DEV_BOARD_B
+#elif defined USE_DEV_BOARD_C
 	BSP_LED_BLU,
 #endif
 	BSP_LED_RED,
@@ -33,4 +33,4 @@ typedef enum {
 /* Exported functions prototypes ---------------------------------------------*/
 
 /* 用于A板时，会无视duty_cycle的值。使用B板时，duty_cycle才有效*/
-int BSP_LED_Set(BSP_LED_Channel_t ch, BSP_LED_Status_t s, int16_t duty_cycle);
+int BSP_LED_Set(BSP_LED_Channel_t ch, BSP_LED_Status_t s, float duty_cycle);
